+++ server/src/api/apiRoutes.ts
@@ -18,6 +18,7 @@ import {
   deleteEnteredGame,
   postEnteredGame,
 } from 'server/features/user/entered-game/enteredGameController';
+import { postSessionRestore } from 'server/features/user/session-restore/sessionRestoreController';
 import {
   getGroup,
   getUser,
@@ -27,7 +28,6 @@ import {
   postLogin,
   postUser,
   postSignup,
-  postRestoreSession,
 } from 'server/features/user/userController';
 import {
   ASSIGNMENT_ENDPOINT,
@@ -38,7 +38,7 @@ import {
   GROUP_ENDPOINT,
   HIDDEN_ENDPOINT,
   LOGIN_ENDPOINT,
-  RESTORE_SESSION_ENDPOINT,
+  SESSION_RESTORE_ENDPOINT,
   RESULTS_ENDPOINT,
   SETTINGS_ENDPOINT,
   SIGNUPTIME_ENDPOINT,
@@ -68,7 +68,7 @@ apiRoutes.post(GROUP_ENDPOINT, postGroup);
 apiRoutes.post(TOGGLE_APP_OPEN_ENDPOINT, postAppOpen);
 apiRoutes.post(ENTERED_GAME_ENDPOINT, postEnteredGame);
 apiRoutes.post(SIGNUP_MESSAGE_ENDPOINT, postSignupMessage);
-apiRoutes.post(RESTORE_SESSION_ENDPOINT, postRestoreSession);
+apiRoutes.post(SESSION_RESTORE_ENDPOINT, postSessionRestore);
 
 /* GET routes */
