+++ server/src/features/user/userController.ts
@@ -10,7 +10,6 @@ import {
   storeGroup,
   login,
   storeSignup,
-  loginWithJwt,
 } from 'server/features/user/userService';
 import { UserGroup } from 'shared/typings/models/user';
 import { isAuthorized } from 'server/utils/authHeader';
@@ -19,7 +18,6 @@ import {
   FAVORITE_ENDPOINT,
   GROUP_ENDPOINT,
   LOGIN_ENDPOINT,
-  RESTORE_SESSION_ENDPOINT,
   SIGNUP_ENDPOINT,
   USERS_BY_SERIAL_ENDPOINT,
   USERS_ENDPOINT,
@@ -30,7 +28,6 @@ import { SaveFavoriteRequest } from 'shared/typings/api/favorite';
 import {
   LoginFormFields,
   RegistrationFormFields,
-  SessionRecoveryRequest,
 } from 'shared/typings/api/login';
 import { sharedConfig } from 'shared/config/sharedConfig';
 import { ConventionType } from 'shared/config/sharedConfig.types';
@@ -74,22 +71,6 @@ export const postLogin = async (
   }
 
   const response = await login(username, password);
-  return res.json(response);
-};
-
-export const postRestoreSession = async (
-  req: Request<{}, {}, SessionRecoveryRequest>,
-  res: Response
-): Promise<Response> => {
-  logger.info(`API call: POST ${RESTORE_SESSION_ENDPOINT}`);
-
-  const { jwt } = req.body;
-
-  if (!jwt) {
-    return res.sendStatus(422);
-  }
-
-  const response = await loginWithJwt(jwt);
   return res.json(response);
 };
