+++ shared/constants/apiEndpoints.ts
@@ -14,4 +14,4 @@ export const SETTINGS_ENDPOINT = '/api/settings';
 export const RESULTS_ENDPOINT = '/api/results';
 export const ENTERED_GAME_ENDPOINT = '/api/enteredgame';
 export const SIGNUP_MESSAGE_ENDPOINT = '/api/signup-info';
-export const RESTORE_SESSION_ENDPOINT = '/api/restore-session';
+export const SESSION_RESTORE_ENDPOINT = '/api/session-restore';
